backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
bitarray==3.4.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
cryptography==45.0.3
cytoolz==1.0.1
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework-api-key==3.1.0
djangorestframework_simplejwt==5.5.0
environ==1.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
factory_boy==3.3.3
Faker==37.4.0
frozenlist==1.6.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
gunicorn==23.0.0
hexbytes==1.3.1
httplib2==0.22.0
idna==3.10
multidict==6.4.4
oauthlib==3.2.2
packaging==25.0
parsimonious==0.10.0
pillow==11.3.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
pyparsing==3.2.3
python-dotenv==1.1.0
pytz==2025.2
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rlp==4.1.0
rsa==4.9.1
sqlparse==0.5.3
toolz==1.0.0
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
web3==7.12.0
websockets==15.0.1
yarl==1.20.0
